Implemented quantization
- Added not transposed weights
- Exported quantized weights
- Fixed make config
- Fixed quantized SIMD
- Implemented int16_t for Apple AMX
- Implemented quantization in export image py
- Implemented quantization in export xl py
- Implemented quantization in tf py
- Implemented shared memory for matmul
- Moved to 10 epochs
- Removed transpose after flatten
- Removed transposing of fc_weights
- Removed unused functions
- Switched to int
- Use transposed weights in io

+++ Source Code/hpp/matrix.hpp
@@ -5,7 +5,7 @@
 
     #ifndef DATA_TYPE
         #ifdef INT
-            typedef int8_t DATA_TYPE;
+            typedef int DATA_TYPE;
         #else
             typedef float DATA_TYPE;
         #endif
